requirements.txt
Django==5.2.1
Pillow==10.3.0
django-environ==0.11.2
pre-commit==4.2.0
boto3==1.38.11
django-storages==1.14.6
django-minio-backend==3.6.0
django-ckeditor==6.2.0
django-modeltranslation==0.19.14
minio==7.2.6
django-mptt==0.17
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
